Penerapan Redux Pada ReactApp (Pemisahan Reducer)
