+++ src/container/Pages/Product/CardProduct/Counter.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import ActionType from '../../../../redux/reducer/globalActionType';
 
 class Counter extends Component {
     // state = {
@@ -49,8 +50,8 @@ const mapStateToProps = (state) => {
 
 const mapDispatchToProps = (dispatch) => {
     return {
-        handlePlus: () => dispatch({ type: 'PLUS_ORDER' }),
-        handleMinus: () => dispatch({ type: 'MINUS_ORDER' })
+        handlePlus: () => dispatch({ type: ActionType.PLUS_ORDER }),
+        handleMinus: () => dispatch({ type: ActionType.MINUS_ORDER })
     }
 }
