+++ src/redux/reducer/globalActionType.js
@@ -1,0 +1,6 @@
+const ActionType = {
+    PLUS_ORDER: 'PLUS_ORDER',
+    MINUS_ORDER: 'MINUS_ORDER'
+}
+
+export default ActionType;
